revert 3c62e25
revert change as it leads to error:

Keyboards require additional configuration to allow for firmware with ZMK Studio enabled. You have attempted to build a keyboard lacking such configuration. Please see https://zmk.dev/docs/features/studio#adding-zmk-studio-support-to-a-keyboard for more information on how to resolve this error, or contact the maintainer of your keyboard\'s firmware for assistance.